Add files via upload
Q1.Check if the frequency of all the digits in a number is same
Input: N = 1234567890
Output: YES
The frequencies of all the digits are same.
i.e. every digit appears same number of times.

Input: N = 1337
Output: NO

Q2.Program to check if a given array is Lucky (all elements are different)
An array is lucky if all digits of the number are different.